chore: move AnalysisResult among other interfaces
